test data output, retrieve with gdb

+++ Inc/hdmi_driver.h
@@ -19,8 +19,8 @@ typedef enum {
 } Camera_StatusTypeDef;
 
 uint8_t BSP_CAMERA_Init(uint32_t Resolution);
-void BSP_CAMERA_ContinuousStart(uint32_t *buff);
-void BSP_CAMERA_SnapshotStart(uint32_t *buff);
+void BSP_CAMERA_ContinuousStart(uint8_t *buff);
+void BSP_CAMERA_SnapshotStart(uint8_t *buff);
 void BSP_CAMERA_Suspend(void);
 void BSP_CAMERA_Resume(void);
 uint8_t BSP_CAMERA_Stop(void);
